ztm-ds-and-algo/src/main/java/bevans/ztm/arrays/MaxSubArray.java
package bevans.ztm.arrays;

public class MaxSubArray {
    public int maxSubArray(int[] nums) {
        var maxSum = Integer.MIN_VALUE;
        var currentSum = 0;

        for (int i = 0; i < nums.length; i++) {
            currentSum += nums[i];
            if (currentSum > maxSum) {
                maxSum = currentSum;
            }

            if (currentSum < 0) {
                currentSum = 0;
            }
        }

        return maxSum;
    }

    public int maxSubArrayCleanCode(int[] nums) {
        int maxSum = nums[0];
        int currentSum = nums[0];

        for (int i = 1; i < nums.length; i++) {
            currentSum = Math.max(nums[i], currentSum + nums[i]);
            maxSum = Math.max(maxSum, currentSum);
        }

        return maxSum;
    }
}
